BlackJack game
BlackJack game
https://replit.com/@kociamadka/BlackJack-game?v=1
A first interesting project from my #100DaysOfCode Challenge. It's an interactive game of BlackJack. You are the player, computer is the dealer. (There are two decks of cards, Aces can be either counted as 1 or 11 points, Jacks/Queens/Kings count as 10, there are no Jokers and each card has the same probability of being picked.)