The Turtle Crossing Game
A game created using the turtle module.
The turtle has to cross the screen (turtle moves with 'Up' and 'Down' arrows on the keyboard) without touching the flying blocks. Every successful crossing takes the user to a next level with increased speed and amount of the flying blocks. Detected collision with any flying block ends with the GAME OVER :))